Move header section to separate component and pass in selectedIds
